requirements.txt
click>=8.0.0
GitPython>=3.1.0
PyGithub>=2.0.0
jinja2>=3.1.0
google-generativeai>=0.3.0